requirements.txt
requests
pytest
selenium
Appium-Python-Client==2.0.0